move openai module
