+++ core/src/transformers/mod.rs
@@ -1,6 +1,5 @@
 pub mod generic;
 pub mod http_handler;
 pub mod ollama;
-pub mod openai;
 pub mod providers;
 pub mod types;
